Moved table init to Engine
